ilPSP: converted external indices to long (2)

+++ src/ilPSP/layer_1.2-ilPSP/ilPSP/IBlockPartitioning_Extension.cs
@@ -74,7 +74,7 @@ namespace ilPSP {
         /// </param>
         /// <returns></returns>
         public static BlockPartitioning GetSubBlocking<T>(this IBlockPartitioning B, T SubvectorIdx, MPI_Comm comm, int FrameBlockSize = -1) 
-            where T : IEnumerable<int>
+            where T : IEnumerable<long>
         {
             List<int[]> Subblk_i0 = new List<int[]>();
             List<int[]> SubblkLen = new List<int[]>();
